- Various cleanups and optimizations:   - MaxValue tracking now done at recording time (predictable branch on cache hitting cell, virtually no slowdown seen).   - Changed from bucket and sub-bucket iteration logic in various routines to flat scanning of the array. - Added support for integerToDoubleValueConversionRatio, prepping for DoubeHistogram support. - Added iterator support for double value tracking - Added setCountAtValue - Started re-work of wire format code (will need to support multiple formats, likely move to helper class).

+++ src/main/java/org/HdrHistogram/LogarithmicIterator.java
@@ -43,7 +43,7 @@ public class LogarithmicIterator extends AbstractHistogramIterator implements It
      * @param valueUnitsInFirstBucket the size (in value units) of the first value bucket step
      * @param logBase the multiplier by which the bucket size is expanded in each iteration step.
      */
-    public LogarithmicIterator(final AbstractHistogram histogram, final int valueUnitsInFirstBucket, final double logBase) {
+    public LogarithmicIterator(final AbstractHistogram histogram, final long valueUnitsInFirstBucket, final double logBase) {
         reset(histogram, valueUnitsInFirstBucket, logBase);
     }
